requirements.txt
# Core dependencies
streamlit==1.30.0
google-generativeai==0.3.0
python-dotenv==1.0.1

# Development dependencies (if needed)
black==23.1.0
pylint==2.17.0
